fixed impossible maps

+++ includes/structs.h
@@ -48,7 +48,7 @@ typedef struct s_player {
 	int		speed_mod;
 	int		has_collisions;
 	int		has_sound;
-	int		on_fire;
+	float		on_fire;
 }	t_player;
 
 typedef struct s_imgptr {
